move generic functions to LPM module
